requirement.txt
Django==1.8.7
six==1.10.0
pytz==2015.4
pillow=2.9.0
uwsgi=2.0.12
django-haystack==2.4.0
django-ckeditor==5.0.2
django-tastypie==0.12.2
django-compressor==1.5
easy-thumbnails==2.2
whoosh==2.7.0
